update import paths after moving auth
